requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
streamlit==1.40.1
numpy==1.26.4

# ── LangChain compatible set (free stack) ─────────────────────────────────
langchain==0.3.25
langchain-core==0.3.63
langchain-community==0.3.24
langchain-text-splitters==0.3.8
langchain-huggingface==0.1.2
langchain-groq==0.2.4

sentence-transformers==3.3.1
torch==2.2.2

# ── Vector Store ──────────────────────────────────────────────────────────
faiss-cpu==1.9.0

# ── PDF Loading ───────────────────────────────────────────────────────────
pypdf==5.1.0
